requirements.txt
streamlit
google-auth
requests
google-generativeai
python-dotenv
